[IT-2979] Improve cost-center tag script
The script to update cost-center tags should update both CostCenter
and CostCenterOther tags on resources. Add logic to update CostCenter
and only update CostCenterOther when the current value for CostCenter
is Other.

Add support for passing in either an EC2 instance ID, or an
arbitrary resource ARN.

Refactoring includes combining _get_cfn_stack_arns and
_get_sc_product_arns into _get_related_arns, and renaming the script
from update_other_tag to update_costcenter_tag.
